+++ packages/block-library/src/post-breadcrumbs/index.php
@@ -1,12 +1,12 @@
 <?php
 /**
- * Server-side rendering of the `core/post-breadcrumb` block.
+ * Server-side rendering of the `core/post-breadcrumbs` block.
  *
  * @package WordPress
  */
 
 /**
- * Renders the `core/post-breadcrumb` block on the server.
+ * Renders the `core/post-breadcrumbs` block on the server.
  *
  * @since 6.8.0
  *
@@ -16,7 +16,7 @@
  *
  * @return string Returns the post breadcrumb for hierarchical post types.
  */
-function render_block_core_post_breadcrumb( $attributes, $content, $block ) {
+function render_block_core_post_breadcrumbs( $attributes, $content, $block ) {
 	if ( ! isset( $block->context['postId'] ) || ! isset( $block->context['postType'] ) ) {
 		return '';
 	}
@@ -88,16 +88,16 @@ function render_block_core_post_breadcrumb( $attributes, $content, $block ) {
 }
 
 /**
- * Registers the `core/post-breadcrumb` block on the server.
+ * Registers the `core/post-breadcrumbs` block on the server.
  *
  * @since 6.8.0
  */
-function register_block_core_post_breadcrumb() {
+function register_block_core_post_breadcrumbs() {
 	register_block_type_from_metadata(
-		__DIR__ . '/post-breadcrumb',
+		__DIR__ . '/post-breadcrumbs',
 		array(
-			'render_callback' => 'render_block_core_post_breadcrumb',
+			'render_callback' => 'render_block_core_post_breadcrumbs',
 		)
 	);
 }
-add_action( 'init', 'register_block_core_post_breadcrumb' );
+add_action( 'init', 'register_block_core_post_breadcrumbs' );
