+++ packages/block-library/src/post-breadcrumbs/style.scss
@@ -1,4 +1,4 @@
-.wp-block-post-breadcrumb {
+.wp-block-post-breadcrumbs {
 	ol {
 		list-style: none;
 		margin: 0;
