+++ packages/block-library/src/style.scss
@@ -38,7 +38,7 @@
 @import "./paragraph/style.scss";
 @import "./post-author/style.scss";
 @import "./post-author-biography/style.scss";
-@import "./post-breadcrumb/style.scss";
+@import "./post-breadcrumbs/style.scss";
 @import "./post-comments-form/style.scss";
 @import "./post-comments-count/style.scss";
 @import "./post-content/style.scss";
