+++ test/integration/fixtures/blocks/core__post-breadcrumbs.html
@@ -1,0 +1,2 @@
+<!-- wp:post-breadcrumbs /-->
+
