+++ test/integration/fixtures/blocks/core__post-breadcrumbs.json
@@ -1,0 +1,11 @@
+[
+	{
+		"name": "core/post-breadcrumbs",
+		"isValid": true,
+		"attributes": {
+			"separator": "/",
+			"showHomeLink": true
+		},
+		"innerBlocks": []
+	}
+]
