+++ test/integration/fixtures/blocks/core__post-breadcrumbs.parsed.json
@@ -1,0 +1,9 @@
+[
+	{
+		"blockName": "core/post-breadcrumbs",
+		"attrs": {},
+		"innerBlocks": [],
+		"innerHTML": "",
+		"innerContent": []
+	}
+]
