+++ test/integration/fixtures/blocks/core__post-breadcrumbs.serialized.html
@@ -1,0 +1,1 @@
+<!-- wp:post-breadcrumbs /-->
